pyproject.toml
[project]
name = "app"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "agno>=2.3.13",
    "fastapi>=0.124.4",
    "groq>=0.37.1",
    "pydantic-settings>=2.12.0",
    "sqlalchemy>=2.0.45",
    "uvicorn>=0.38.0",
]
